+++ spec/support/sample_classes/sample_commands.rb
@@ -3,34 +3,34 @@ end
 
 module SampleCommands
   class SuccessCommand < AuxiliaryRails::AbstractCommand
-    def call
+    def perform
       success!
     end
   end
 
   class DoubleStatusSetCommand < AuxiliaryRails::AbstractCommand
-    def call
+    def perform
       failure!
       success!
     end
   end
 
   class SuccessWithErrorsCommand < AuxiliaryRails::AbstractCommand
-    def call
+    def perform
       errors.add(:command, :error)
       success!
     end
   end
 
   class FailureWithErrorsCommand < AuxiliaryRails::AbstractCommand
-    def call
+    def perform
       errors.add(:command, :test_failure_message)
       failure!
     end
   end
 
   class ValidationErrorsCommand < AuxiliaryRails::AbstractCommand
-    param :age
+    argument :age
 
     validates :age,
       numericality: {
@@ -38,9 +38,7 @@ module SampleCommands
         greater_than: 18
       }
 
-    def call
-      return failure! unless valid?
-
+    def perform
       success!
     end
   end